requirements.txt
torch>=2.6.0
torchvision>=0.21.0
timm
einops
accelerate
transformers>=4.51.3
diffusers
opencv-python-headless
scipy
wandb
matplotlib
Pillow
tqdm
omegaconf
python-dotenv
ninja
ipykernel

